split jwt cli out into workspace member
