+++ jwt_bearer/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "jwt-bearer"
+version = "0.1.0"
+edition = "2021"
+license = "Apache-2.0"
+
+[[bin]]
+name = "jwt-bearer"
+path = "src/main.rs"
+
+[dependencies]
+appcheck-backend = { path = "../" }
+bs58 = { workspace = true }
+clap = { version = "4", features = ["derive"]}
+coarsetime = "*"
+jwt-simple = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
